% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctn0094.oliven-package.R
\docType{package}
\name{ctn0094.oliven-package}
\alias{ctn0094.oliven}
\alias{ctn0094.oliven-package}
\title{ctn0094.oliven: Analysis of CTN-0094 Data}
\description{
This is an analysis that predicts treatment response for people undergoing care for Opioid Use Disorder (OUD).
}
\author{
\strong{Maintainer}: Nat Oliven \email{oliven@umich.edu} (\href{https://orcid.org/0009-0005-2359-319X}{ORCID})

Authors:
\itemize{
  \item Nat Oliven \email{oliven@umich.edu} (\href{https://orcid.org/0009-0005-2359-319X}{ORCID})
}

}
\keyword{internal}
